+++ src/Commands/ValidatorsCommands.php
@@ -80,4 +80,24 @@ class ValidatorsCommands extends DrushCommands
 
         $this->validateBundle($entityType, $bundle);
     }
+
+    protected function validateBundle(string $entityTypeId, string $bundle): void
+    {
+        $entityTypeDefinition = $this->entityTypeManager->getDefinition($entityTypeId);
+
+        if ($entityTypeDefinition && $bundleEntityType = $entityTypeDefinition->getBundleEntityType()) {
+            $bundleDefinition = $this->entityTypeManager
+                ->getStorage($bundleEntityType)
+                ->load($bundle);
+        }
+
+        if (!isset($bundleDefinition)) {
+            throw new \InvalidArgumentException(
+                t("Bundle ':bundle' does not exist on entity type with id ':entityType'.", [
+                    ':bundle' => $bundle,
+                    ':entityType' => $entityTypeId,
+                ])
+            );
+        }
+    }
 }