+++ src/Commands/ValidatorsTrait.php
@@ -30,24 +30,4 @@ trait ValidatorsTrait
             );
         }
     }
-
-    public function validateBundle(string $entityTypeId, string $bundle): void
-    {
-        $entityTypeDefinition = $this->entityTypeManager->getDefinition($entityTypeId);
-
-        if ($entityTypeDefinition && $bundleEntityType = $entityTypeDefinition->getBundleEntityType()) {
-            $bundleDefinition = $this->entityTypeManager
-                ->getStorage($bundleEntityType)
-                ->load($bundle);
-        }
-
-        if (!isset($bundleDefinition)) {
-            throw new \InvalidArgumentException(
-                t("Bundle ':bundle' does not exist on entity type with id ':entityType'.", [
-                    ':bundle' => $bundle,
-                    ':entityType' => $entityTypeId,
-                ])
-            );
-        }
-    }
 }