+++ src/Commands/WmModelCommands.php
@@ -10,15 +10,15 @@ use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\wmmodel\Factory\ModelFactory;
 use Drupal\wmscaffold\Service\Generator\ModelClassGenerator;
 use Drush\Commands\DrushCommands;
+use Drush\Drupal\Commands\core\EntityTypeBundleAskTrait;
 use PhpParser\PrettyPrinter\Standard as PrettyPrinter;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Filesystem\Filesystem;
 
 class WmModelCommands extends DrushCommands implements SiteAliasManagerAwareInterface
 {
-    use AskBundleTrait;
-    use QuestionTrait;
     use RunCommandTrait;
+    use EntityTypeBundleAskTrait;
 
     /** @var EntityTypeManagerInterface */
     protected $entityTypeManager;