+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Drush 11 dependency
 
 ### Removed
-- Remove `field-create` since it has been added to Drush in v11
+- Remove field commands since they have been added to Drush in v11
+- Remove `AskBundleTrait`
+- Remove `QuestionTrait` and `ChoiceQuestion`
 
 ## [1.12.1] - 2021-12-12
 ### Fixed